Rewrite Groovy Kotlin DSL tests in Kotlin
